Flexible panel layout init
