+++ src/app/App.module.css
@@ -1,0 +1,4 @@
+.main {
+  height: 100vh;
+  width: 100%;
+}
