+++ src/features/map/Map.css
@@ -1,6 +1,6 @@
 .Map {
-  width: 100vw;
-  height: 100vh;
+  width: 100%;
+  height: 100%;
 }
 
 .Map--NavigationControl {
