+++ src/features/widgets/MapLegend.css
