+++ src/features/widgets/MapLegend.js
@@ -12,8 +12,8 @@ export function MapLegend({ clusters }) {
   )
 
   return (
-    <div className="Map--Overlays--Box MapLegend">
-      <fieldset className="field">
+    <div className="MapLegend">
+      <fieldset className="field" style={{ maxWidth: 300 }}>
         <label className="label" style={{ marginBottom: 18 }}>
           Number of Typologies: <span className="tag">{numberOfClusters}</span>
         </label>
