+++ src/features/widgets/WidgetBox.js
@@ -1,15 +1,6 @@
-import { X as CloseIcon } from 'react-feather'
-
-import { IconButton } from '../../common/IconButton'
-
 export function WidgetBox({ children, onClose, ...props }) {
   return (
     <div className="Widgets--Box" {...props}>
-      <IconButton
-        Icon={CloseIcon}
-        className="Widgets--Box--CloseButton"
-        onClick={onClose}
-      />
       {children}
     </div>
   )
