+++ src/index.css
@@ -18,6 +18,7 @@
   --color-highlight-c: var(--orange);
   --color-light-grey: var(--light-grey);
   --color-mid-grey: var(--mid-grey);
+  --color-dark-grey: var(--dark-grey);
   --color-background: var(--white);
   --color-code: var(--light-black);
